Add custom corners support in refit method
The refit method ignored any corner index array that was passed to
the main function. This implements support for providing custom
corner indices.

Similar to how corners can be provided to `curve_fit_cubic_to_points`,
this adds the same functionality to `curve_fit_cubic_to_points_refit`.

The caller can pass in an array of indices corresponding to the points
that should be treated as corners. Corners will be part of the output
curve and use free handles.

Also makes sure that the `r_corner_index_array` is correctly written.